Branch to work on Call Completion Supplementary Services in Asterisk.
What are the Call Completion Supplementary Services? They are
Call Completion Busy Subscriber(CCBS) and Call Completion
No Response(CCNR).

This branch will add the support for these features to Asterisk.
This will be accomplished through "generic" methods which will
work for any technology, as well as protocol-specific methods.
The current plan is to support ISDN methods as well as SIP for
the protocol-specific methods.

After the typical automerge commit, the first item that will
be added to this branch will be an architecture document describing
at a relatively high level how the service will work within
Asterisk.