Add some logic to help when parsing channel driver config files.
Also added the necessary logic to chan_sip.c to parse config files.
Next step is to get rid of all "ccss" from public function names
in favor of "cc." The mixture is already annoying even in this
early stage, and it seems like the best idea to go with the
shorter name.