send dummy SIP packet with metadata on startup/end
On stream startup send a SIP INVITE message and on stream end send a BYE
message.

Each of those include practically the same data and are intended to pass
information about the stream to the remote party. The remote recording
server is not expected to respond to those messages.

This method is used to pass CDR information to the remote side.
Currrently
only src, dst, and clid are passed along.

Also note that unlike any standard SIP call, a separate INVITE/BYE pair
will be used for any separate direction (tx, rx) of the channel.