Flesh out the skeleton for translating an interface tree into monitor structs a bit.
There are bits left as comments because I haven't yet determined the
best way to find if a monitor for a particular extension/device already
exists. My thinking is that while the graph can ceratainly be searched
using the common DFS and BFS methods, doing so will be incredibly inefficient
and the coding for doing so could be painful, especially if someone figures
out a way to put a loop in the graph. So, instead, I believe what I'll do
is to have the monitors in an ao2_container so that I can easily do an O(1)
lookup when I need to. The tradeoff is that the refcount logic will need
a bit more hand holding, but I'd much rather have that.