Checking in progress with monitor creation.
Most of what's necessary is in place, but lord does
it need lots and lots and lots of tweaking. One
big thing I need to do next is to make sure that
the frame data in the AST_CONTROL_CC frame has
all that is necessary so that the ensuing
ast_cc_interface has the information so that the
monitor can then have the information.

I think I'll write a specialized frame queuing
function for the frame type.