Adding to the janitor list.
For new readers: The janitor list is a list of tasks we need help with in the Asterisk project. Taking up 
one of these is often a good way to get into Asterisk development and getting a lot of developers in 
the project to be grateful. It's stuff we could spend time on when the bug tracker is empty, when our
employers hasn't filled our task lists and our servers is running bugfree and happily without any issues.

If you want to start working on one of these small projects, feel free to ask for help in the #asterisk-dev
channel on IRC or asterisk-dev mailing list. We'll be more than happy to help you to start and reach
goal.

Thank you for your help.

</end of long commit message>