Move dial_cc_interfaces datastore out of app_dial and into ccss.c
This commit is pretty much a giant move operation to put some code that
was only in app_dial into ccss.c. There are two really good reasons for
doing this:

1. If it were ever decided that more than just the Dial application should
allow for CC operations, then having a central location for the datastore
and a set of API calls to get and set data makes the task much easier.

2. More importantly, the addition of API calls to interact with the 
datastore allows for other files outside of app_dial to get and
set items. This allows for, say, chan_local to set data before Dial is
even called so that we can know if things like "/nb" are at the end
of the channel name. It also will allow for called channel drivers to
query what the core ID of the call for which they have just queued a
CC frame will be. This will require a bit more modification, actually.

The next commit will be to fix up the comments in the moved code, change
some function names so they make more sense, and doxygenify all new API
calls I've added in this commit.