Change when core_id is allocated and write API call to retrieve it.
So, let's say that you operate a channel driver, and you decide
that you can queue an AST_CONTROL_CC frame because the outbound
channel can handle CC requests. Well, when the time comes to
start monitoring that device, all you'll be given to identify
the call is a core_id. So, now at the time that you queue
the CC frame, the channel driver may use the API call to store
the core_id which will be used when monitoring time comes.