Change ast_setup_cc_recall_datastore to take a core_id instead of an agent.
The change is made for two reasons.
1) This should make it easier for channel drivers to use.
2) This avoids a potential race condition. This function is most likely to
be called in a channel driver monitor thread. Since agents are not refcounted,
it is potentially dangerous to get a pointer to an agent and expect it to stay
valid since the CC thread could destroy the agent out from under the channel
monitor thread. By using the core_id, we can guarantee that the agent is accessed
through a cc_core_instance, which is refcounted, which makes the operation safe.