Change the model by which core instances are created.
It used to be that all outbound devices were collected
together and then a core instance was created after the
call had either been answered or hung up.

Since ISDN requires that the calling channel driver be
made aware of the ability to offer CC prior to sending
the ALERTING message or DISCONNECT message, it is necessary
to have the agent created for the caller as soon as possible.
Now we create a core instance and an agent as soon as we
read the first CC control frame.

This change touches quite a few areas in addition to
what was mentioned above.

1. ast_cc_core_init_instance is now a private function
called cc_core_init_instance.
2. ast_cc_create_new_core has been removed since there
is no reason for anything outside of ccss to request for
a CC core to be created.
3. cc_core_init_instance now returns a reference to a
core instance instead of the core_id of the newly
created core_instance.