Change the way in which a recall is handled.
Now, when a callee becomes available, the core does
not issue a status request. The reason for this is that
a status request is not always something that can be guaranteed
to return quickly. Instead, when the callee becomes available,
the core alerts the agent of such, and it is assumed that the
agent will notify the caller that he may attempt to recall. If
the agent is unable to reach the caller, or if the agent has
the ability to check beforehand that the caller is not going
to be available, it can then signal to the core that the caller
is unavailable and abort the recall attempt.

The other change is that the CC_RECALLING state is actually entered
now when the caller actually initiates the recall.