Remove ast_cc_get_recall_core_id and replace with ast_cc_is_recall.
The reason for this is that the typical use case for this is when
a channel is called, the channel driver will wish to know if the
call to this channel is a recall or not.

This new function works by checking to see if the channel in question
exists in the list of originally-called channels and that it is
being called from the same extension and context as it was originally.
If not, then this channel has perhaps been added to the CC_INTERFACES
channel variable by the admin, or the call has been routed to a completely
different device for whatever reason.

The functionality isn't perfect yet, however. For instance, if Dial is
called twice, the second time that Dial is called, the ast_cc_is_recall
function will still return non-zero. This last portion will be taken
care of in the next commit.