Change CC complete timing.
One situation that rmudgett and I discussed is one in which
CC is declared to be complete before a downstream endpoint has received
the SETUP/INVITE/whatever for the recall. It is possible, though
not likely, that the CC structures could be destroyed prior to
the recall arriving on a downstream server, thus resulting in
the callee not being properly dialed. If that didn't make sense,
sorry.

Anyway, the remedy for this is to not declare CC to be completed
when the first frame arrives, but rather to wait until the end of
the wait_for_answer function in Dial, so we have a final verdict on
the call.

At this point, there are still some compilation errors in ccss.c, all
of which are due to the fact that the cc_recall datastore definition
is further down in the file than it should be.

The next steps are to get those errors fixed, do another run-through
to get all doxygen up-to-date and accurate, and then re-run all my
tests to be sure that I have not introduced any new bugs while fixing
the problems being addressed the last few days.