Fix a bug that affected calls to local channels.
One change that was made was to allocate a cc_core_instance
and an ast_cc_agent when we read the first AST_CONTROL_CC frame
we receive. The problem is that if a call went through a local
channel, then the structures would be allocated on the local channel
instead of the original caller.

To fix the problem, I changed the is_cc_offerable field  of the
cc_dialed_interfaces structure to be is_original_caller instead, since
that was what it was used for anyway. Then, I use this field's value
to determine if the inbound channel to Dial is the original caller.
If it is not, then the assumption is made that the call is from a local
channel instead, and so we indicate the condition to the local channel.

chan_local is cool enough to automatically queue the frame over to the
other side for us, so it all works out in the end and everyone's happy.
God bless America.