Create a bottom-up monitor destruction mechanism, as well as a public function to trigger it.
When a specific monitor fails for some reason, it may call ast_cc_monitor_failed in
order to start the bottom-up destruction process, which works essentially in the opposite
direction of prune_links. This way, we destroy the individual faulty monitor instead of
destroying everything just because one has failed. The logic works such that the failure
may propagate up the tree if necessary.

The next step will be to analyze my usage of ast_cc_failed in the core to be ensure that
I should not actually be using ast_cc_monitor_failed in those circumstances.