A couple of changes for ast_cc_monitor_failed taskprocessor callback.
The logic was just garbage before. it should work as it is supposed to now.
In addition, the cc_monitor_failed internal function will check if the current
state is CC_RECALLING. If it is, then the monitor failure will be ignored since
the call will result in the proper destructors being called.