Fix a logic error in cc_extension_monitor_request_cc.
The old logic would see if the list of child monitors was
empty in order to determine the return value. This is not
correct, though, since an extension monitor may be involved
in multiple calls. Instead, we need to return failure if
all of the child monitors on whom we called the request_cc
callback failed.