Add method to search for an agent given a piece of data.
While trying to write code to handle an incoming PUBLISH, it became
clear that the way I would need to find the appropriate CC agent was
based on the request URI of the PUBLISH. I needed to match the URI
to the stored URI in the agent's private_data. Unfortunately, I can't
search the container of cc_core_instances in chan_sip because the container
and the cc_core_instance struct are both private to ccss.c. However, it
was not too difficult to create a public API call which could use ao2_callback
within ccss.c. This is very Matryoshka-esque code.