Add a new channel tech callback to call a callback for CC purposes.
When attempting to request an analog channel, or perhaps when trying
to request an ISDN channel when no B channels are available, ast_request
will fail to return a channel to the requester. The result of this is
that

1) We can't figure out the CC configuration parameters for the channel that
would have been returned.
2) We can't read a CC control frame.

In order to alleviate the problem, a new channel tech function has been created,
cc_callback. What this does is it allows us to dive into the channel driver, and
find out the interface that would have been found had ast_request not failed in
the first place. Once we find such an interface, we can call a supplied callback
so that the CC core can do whatever processing it may want or need to do.

It's a bit convoluted. Unfortunately, chan_dahdi just doesn't behave the same
way as most other channel drivers, and so special considerations have to be made.

Right now, the code in here does not compile because the chan_dahdi CC configuration
parsing is not implemented. That's up next, and then will come the fun part where
we attempt to figure out how to handle the situation in the CC core.