Add another oddball monitor API/agent callback pair. This time, "party_b_free"
This is something that really is only needed when Asterisk acts as the NT side
for an ISDN PTMP caller and as the TE side for an ISDN PTMP callee (i.e. the
ISDN network). Since this requires receiving a message and passing it through
to the other side, it is implemented as a core API and agent callback. Plus, you
never know when a future supported protocol might need something similar.