Modify dahdi_read a bit.
The logic needs to operate whether the call_forward string is set or not. It
will work either way. Don't set needringing to 0 if we haven't sent back a
ringing frame. Add XXX note to be sure to implement a way of telling if we
already sent a CC frame back so we don't keep sending them forever.