Convert the monitor_instance to use string fields and add a new peername stringfield.
The stringfield change makes sense because it could potentially save
a bit of space and the strings will likely only be set once or twice
during a typical CC transaction.

The peername part is being added as a means of finding a specific
sip_monitor_instance. When it comes time to transmit a SUBSCRIBE or
PUBLISH having to do with CC, we'll need to send to the appropriate
URI, but we'll need to get address information based on the peer with
which that URI is associated.