Use refcounting for agents instead of straight allocation.
The reason for this is that there are times where an entity needs
to get an ast_cc_agent structure. In order to do this, he must use
the ast_cc_agent_callback function to find the appropriate agent. The
problem is that since the agent is a component of the cc_core_instance
structure, and cc_core_instance is internal to ccss.c, it become impossible
to properly keep the refcount of the cc_core_instance correct.

The solution instead is to also keep a refcount of ast_cc_agents. This way,
when ast_cc_agent_callback is called, we can give a reference of the agent
to the caller while not having to worry about the refcount of the cc_core_instance.