Stop monitor CC available routines from trying to figure out the monitor type.
cc_build_payload() was trying to figure out what monitor type to use based
upon the monitor policy and the channel name.  Not a good thing.  So I
made it take a monitor type parameter.  Most of the other changes are
ripple effects of this change.  In the process, I fixed a bunch of
failure conditions and added more doxygen comments.