Place locks around operations on the monitor list.
In most cases, this probably isn't necessary, but I would like
to make sure that the core doesn't try to mess with a list while
a channel driver is traversing it to set up the CC_INTERFACES
channel variable or some such. Most of this is being overly
protective, but it'll be all right.

I also messed a bit with the locking of the channel and reference
counts of the recall data interface list when setting the CC_INTERFACES
channel variable.