Create new configuration option, cc_agent_dialstring.
With this option, one may specify a dialstring to be used
when re-calling the original caller when generic call completion
is used.

See the notes in ccss.conf.sample for more information on the
option. Like all CC configuration options, this may be set
using the CALLCOMPLETION dialplan function as well.