Add some missing code regarding CC EPA entries.
First, I never set the instance_data field of the call_completion
EPA entry when I allocated it, so I needed to do that. This allowed
me to not segfault when an error occurred.

Next, I added a destructor callback to EPA entry static_data
members. This way, the destructor is called properly when
the EPA entry is destroyed.

There is still a bit of murkiness to clean up with regards
to multiple endpoints being dialed, but things appear to work
wonderfully when dialing a single endpoint.