Fix a race condition in generic device monitoring.
There was a race condition that could happen where a device
state event would fire between when we destroyed our generic
monitor and when we unsubscribed from device state. The problem
was that the device state event would result in incrementing and
decrementing an invalid reference to the freed monitor.

I have addressed this by having the device state callback not
have a pointer to the monitor, and instead has just the name
of the monitor. This string is not freed until after the
event is unsubscribed, thus averting the race condition.