Add a case for AST_OPTION_DEVICE_NAME to sip_queryoption.
I found that when using call completion to place a call to
two distinct sip endpoints using a connection to the a single
peer, i.e. Dial(SIP/1000@asterisk&SIP/2000@asterisk) where
"asterisk" is a peer in sip.conf, call completion was not
working as expected. The problem was that both channels are
named SIP/asterisk-<seqno>. Since the call completion core
will drop any CC frames received from what it perceives to
be the same device, only the first call leg over which we
queued a CC frame would be eligible for requesting call
completion.

The solution to this is to fill in the AST_OPTION_DEVICE_NAME
portion of chan_sip's sip_queryoption. This way, whenever
ast_channel_get_device_name is called on a SIP channel, a more
exact device name will be returned. In the sample above, the
two devices returned will be SIP/1000@asterisk and SIP/2000@asterisk.
In the case where a local SIP peer is dialed, the current scheme
can result in a somewhat odd name like SIP/1000@1000. However, this
causes no harmful side effects so for now anyway, there's no real
need to change this.