Add unmonitored extensions to CC_INTERFACES chanvar.
This is one of the last major missing pieces for CCSS. Prior
to this change, only monitored devices were added to the
CC_INTERFACES channel variable during a recall. However, our
design was for all originally-dialed devices to be represented
in this variable. This set of changes allows for that to occur.
The comments in the code, especially for the extension_child_dialstring
structure should explain how this works pretty well.

The next major item for me to look at is going to be lock order. The
monitor restructuring that was done several weeks ago added locking
where previously there was none. I'm going to look for specific areas
where lock order may be an issue.