Change from doubly-linked lists to singly-linked lists.
The change made two commits ago to allow for all devices to be added to
the CC_INTERFACES channel variable changed how list traversal is done
when making monitor callbacks. The result is that there is no longer
a good reason to be using doubly-linked lists. Therefore, they have
been removed. The backwards traversals are all forward traversals now.