Merged revision 247904 from https://origsvn.digium.com/svn/asterisk/be/branches/C.2-...
..........
r247904 | rmudgett | 2010-02-19 10:49:44 -0600 (Fri, 19 Feb 2010) | 49 lines

Make chan_misdn DTMF processing consistent with other channel technologies.

The processing of DTMF tones on the receiving side of an ISDN channel is
inconsistent with the way it is handled in other channels, especially
DAHDI analog.  This causes DTMF tones sent from an ISDN phone to be
doubled at the connected party.

We are using the following 2 options of misdn.conf
1) astdtmf=yes
2) senddtmf=yes

Option one is necessary because the asterisk DSP DTMF detection is better
than mISDN's internal DSP.  Not as many false positives.

Option two is necessary to transmit DTMF tones end to end when mISDN
channels are connected to SIP channels with out of band DTMF for example.

The symptom is that DTMF tones sent by an ISDN phone are doubled on the
way through asterisk when two mISDN channels are connected with a Local
channel in between or if it is bridged to an analog channel.

The doubling of DTMF tones is because DTMF is passed inband to asterisk by
the mISDN channel and passed out of band once again after the release of
the DTMF tone.  Passing it inband is wrong.  Neither an analog channel nor
SIP channel passes DTMF inband if configured to inband DTMF.  Analog and
SIP channels filter out the DTMF tones because they use the voice frames
returned by ast_dsp_process.  But chan_misdn passes the unfiltered input
voice frames instead.

To overcome one aspect of the problem, the doubling of DTMF tones when two
mISDN channels are directly bridged, someone made an 'optimization', where
in that case the DTMF tone passed out-of-band to the peer channel is not
translated to an inband tone at the transmit side.  This optimization is
bad because it does not work in general.  For example, analog channels or
mISDN channels when bridged through an intermediary local channel will
generate DTMF tones from out-of-band information.  Also, of course, it
must not be done when there is no inband DTMF available.

This patch fixes the issue.  Now chan_misdn will filter the received
inband DTMF signal the same as other channel types.

Another change included: No need to build an extra translation path
because ast_process_dsp does it if required.

Patches:
	misdn-dtmf.patch

JIRA ABE-2080