Add a new CDR backend registration method and adapt the backends that need it.
The CDR core now allows a CDR module to register multiple "sinks".  That way,
for a module that allows posting of records to multiple destinations, a
failure for posting to one destination can be handled specifically such that
a retry is only made for that specific destination, and not for the whole
CDR backend.  With that in place, the CDR core can ensure that every CDR
record is successfully logged exactly once to every intended destination.

I have also done some manual testing and things are working as expected.  I
still want to devise some good automated testing for this stuff, though ...