1) Rework HAVE_PRI_SERVICE_MESSAGES to not use the active values directly from the database.  Database access is likely expensive.  Database access now only happens on initialization, destruction, and when the B channel is taken in or out of service.  (This change is not related to call waiting but it would cause the search for a call waiting interface to be very expensive and slow down D channel message servicing.)
2) Change chan_dahdi channel search to "find" a call waiting interface if
no B channels are available on a span that has call waiting enabled.

3) Sorted the span controller channel private pointer array just like the
chan_dahdi interface list.  B channels are now found in the expected order
for incoming calls if the channels were declared out of order in
chan_dahdi.conf.  (This change is no longer required for call waiting.)